requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-docx==1.1.2
PyPDF2==3.0.1
openai==1.59.3
anthropic==0.42.0
pydantic==2.10.4
python-dotenv==1.0.1
redis[hiredis]==5.2.1
